+++ src/containers/LedgerHoc/constants.js
@@ -1,0 +1,21 @@
+/*
+ *
+ * LedgerHoc constants
+ *
+ */
+
+export const LEDGER_CONNECTED = 'src/LedgerHoc/LEDGER_CONNECTED';
+export const LEDGER_DISCONNECTED = 'src/LedgerHoc/LEDGER_DISCONNECTED';
+export const LEDGER_ETH_CONNECTED = 'src/LedgerHoc/LEDGER_ETH_CONNECTED';
+export const LEDGER_ETH_DISCONNECTED = 'src/LedgerHoc/LEDGER_ETH_DISCONNECTED';
+export const LEDGER_ERROR = 'src/LedgerHoc/LEDGER_ERROR';
+
+export const FETCH_LEDGER_ADDRESSES = 'src/LedgerHoc/FETCH_LEDGER_ADDRESSES';
+export const FETCHED_LEDGER_ADDRESS = 'src/LedgerHoc/FETCHED_LEDGER_ADDRESS';
+export const SAVE_LEDGER_ADDRESS = 'src/LedgerHoc/SAVE_LEDGER_ADDRESS';
+
+export const LEDGER_CONFIRM_TX_ON_DEVICE = 'src/LedgerHoc/LEDGER_CONFIRM_TX_ON_DEVICE';
+export const LEDGER_CONFIRM_TX_ON_DEVICE_DONE = 'src/LedgerHoc/LEDGER_CONFIRM_TX_ON_DEVICE_DONE';
+
+export const INIT_LEDGER = 'src/LedgerHoc/INIT_LEDGER';
+export const DEFAULT_ACTION = 'src/LedgerHoc/DEFAULT_ACTION';
