+++ src/containers/LedgerHoc/reducer.js
@@ -1,0 +1,23 @@
+/*
+ *
+ * LedgerHoc reducer
+ *
+ */
+
+import { fromJS } from 'immutable';
+import {
+  DEFAULT_ACTION,
+} from './constants';
+
+const initialState = fromJS({});
+
+function ledgerHocReducer(state = initialState, action) {
+  switch (action.type) {
+    case DEFAULT_ACTION:
+      return state;
+    default:
+      return state;
+  }
+}
+
+export default ledgerHocReducer;
