+++ src/containers/LedgerHoc/selectors.js
@@ -1,0 +1,25 @@
+import { createSelector } from 'reselect';
+
+/**
+ * Direct selector to the ledgerHoc state domain
+ */
+const selectLedgerHocDomain = (state) => state.get('ledgerHoc');
+
+/**
+ * Other specific selectors
+ */
+
+
+/**
+ * Default selector used by LedgerHoc
+ */
+
+const makeSelectLedgerHoc = () => createSelector(
+  selectLedgerHocDomain,
+  (substate) => substate.toJS()
+);
+
+export default makeSelectLedgerHoc;
+export {
+  selectLedgerHocDomain,
+};
