+++ src/containers/WalletHOC/constants.js
@@ -48,19 +48,6 @@ export const TRANSFER_ERC20 = 'src/WalletHOC/TRANSFER_ERC20';
 export const TRANSFER_SUCCESS = 'src/WalletHOC/TRANSFER_SUCCESS';
 export const TRANSFER_ERROR = 'src/WalletHOC/TRANSFER_ERROR';
 
-// Ledger Nano S constants
-export const LEDGER_CONNECTED = 'src/WalletHOC/LEDGER_CONNECTED';
-export const LEDGER_DISCONNECTED = 'src/WalletHOC/LEDGER_DISCONNECTED';
-export const LEDGER_ETH_CONNECTED = 'src/WalletHOC/LEDGER_ETH_CONNECTED';
-export const LEDGER_ETH_DISCONNECTED = 'src/WalletHOC/LEDGER_ETH_DISCONNECTED';
-export const FETCH_LEDGER_ADDRESSES = 'src/WalletHOC/FETCH_LEDGER_ADDRESSES';
-export const FETCHED_LEDGER_ADDRESS = 'src/WalletHOC/FETCHED_LEDGER_ADDRESS';
-export const SAVE_LEDGER_ADDRESS = 'src/WalletHOC/SAVE_LEDGER_ADDRESS';
-export const INIT_LEDGER = 'src/WalletHOC/INIT_LEDGER';
-export const LEDGER_ERROR = 'src/WalletHOC/LEDGER_ERROR';
-export const LEDGER_CONFIRM_TX_ON_DEVICE = 'src/WalletHOC/LEDGER_CONFIRM_TX_ON_DEVICE';
-export const LEDGER_CONFIRM_TX_ON_DEVICE_DONE = 'src/WalletHOC/LEDGER_CONFIRM_TX_ON_DEVICE_DONE';
-
 // Trezor constants
 export const TREZOR_CONNECTED = 'src/WalletHOC/TREZOR_CONNECTED';
 export const TREZOR_DISCONNECTED = 'src/WalletHOC/TREZOR_DISCONNECTED';
