+++ src/containers/WalletHOC/reducer.js
@@ -10,6 +10,19 @@ import abiDecoder from 'abi-decoder';
 
 import { CHANGE_NETWORK } from 'containers/App/constants';
 
+
+import {
+  FETCHED_LEDGER_ADDRESS,
+  LEDGER_CONNECTED,
+  LEDGER_DISCONNECTED,
+  LEDGER_ETH_CONNECTED,
+  LEDGER_ETH_DISCONNECTED,
+  LEDGER_ERROR,
+  SAVE_LEDGER_ADDRESS,
+  LEDGER_CONFIRM_TX_ON_DEVICE,
+  LEDGER_CONFIRM_TX_ON_DEVICE_DONE,
+} from 'containers/LedgerHoc/constants';
+
 import {
   LOAD_WALLET_BALANCES,
   LOAD_WALLET_BALANCES_SUCCESS,
@@ -33,13 +46,6 @@ import {
   TRANSFER,
   TRANSFER_SUCCESS,
   TRANSFER_ERROR,
-  LEDGER_CONNECTED,
-  LEDGER_DISCONNECTED,
-  LEDGER_ETH_CONNECTED,
-  LEDGER_ETH_DISCONNECTED,
-  LEDGER_ERROR,
-  FETCHED_LEDGER_ADDRESS,
-  SAVE_LEDGER_ADDRESS,
   TREZOR_CONNECTED,
   TREZOR_DISCONNECTED,
   FETCHED_TREZOR_ADDRESS,
@@ -49,8 +55,6 @@ import {
   LOAD_TRANSACTIONS_ERROR,
   LOAD_BLOCK_HEIGHT_ERROR,
   LOAD_BLOCK_HEIGHT_SUCCESS,
-  LEDGER_CONFIRM_TX_ON_DEVICE,
-  LEDGER_CONFIRM_TX_ON_DEVICE_DONE,
   TREZOR_CONFIRM_TX_ON_DEVICE,
   TREZOR_CONFIRM_TX_ON_DEVICE_DONE,
 } from './constants';
