+++ src/containers/WalletHOC/tests/actions.test.js
@@ -5,12 +5,8 @@ import {
   decryptWallet,
   decryptWalletFailed,
   decryptWalletSuccess,
-  ledgerEthAppConnected,
-  ledgerError,
   deleteWallet,
-  saveLedgerAddress,
   showDecryptWalletModal,
-  initLedger,
   loadWalletBalances,
   loadWalletBalancesSuccess,
   loadWalletBalancesError,
@@ -24,6 +20,7 @@ import {
   loadPricesSuccess,
   loadPricesError,
 } from '../actions';
+
 import {
   CREATE_WALLET_FROM_MNEMONIC,
   CREATE_WALLET_FAILURE,
@@ -31,12 +28,9 @@ import {
   DECRYPT_WALLET,
   DECRYPT_WALLET_FAILURE,
   DECRYPT_WALLET_SUCCESS,
-  LEDGER_ETH_CONNECTED,
-  LEDGER_ERROR,
   DELETE_WALLET,
   SHOW_DECRYPT_WALLET_MODAL,
   TRANSFER,
-  INIT_LEDGER,
   LOAD_WALLET_BALANCES,
   LOAD_WALLET_BALANCES_SUCCESS,
   LOAD_WALLET_BALANCES_ERROR,
@@ -51,7 +45,6 @@ import {
   LOAD_PRICES_ERROR,
 } from '../constants';
 
-import getFriendlyError from '../../../utils/friendlyErrors';
 
 describe('WalletHoc actions', () => {
   describe('deleteWallet Action', () => {
@@ -292,62 +285,6 @@ describe('WalletHoc actions', () => {
         error,
       };
       expect(decryptWalletFailed(error)).toEqual(expected);
-    });
-  });
-
-  describe('initLedger Action', () => {
-    it('returns expected output', () => {
-      const expected = {
-        type: INIT_LEDGER,
-      };
-      expect(initLedger()).toEqual(expected);
-    });
-  });
-
-  describe('ledgerEthAppConnected Action', () => {
-    it('returns expected output', () => {
-      const id = '048ncjdh39';
-      const descriptor = 'desc';
-      const expected = {
-        type: LEDGER_ETH_CONNECTED,
-        id,
-        descriptor,
-      };
-      expect(ledgerEthAppConnected(descriptor, id)).toEqual(expected);
-    });
-  });
-
-  // should dispatch new add wallet action
-  describe('saveLedgerAddress Action', () => {
-    it('returns expected output', () => {
-      const deviceId = '0x0028342093';
-      const address = '0x000';
-      const name = 'wallet12';
-      const derivationPath = 'm/0/0123';
-      const newWallet = {
-        deviceId,
-        address,
-        type: 'lns',
-        name,
-        derivationPath,
-      };
-      const expected = {
-        type: CREATE_WALLET_SUCCESS,
-        newWallet,
-      };
-      expect(saveLedgerAddress(name, derivationPath, deviceId, address)).toEqual(expected);
-    });
-  });
-
-  describe('ledgerError Action', () => {
-    it('converts error and returns expected output', () => {
-      const error = { id: 'ListenTimeout' };
-      const friendlyError = getFriendlyError(error);
-      const expected = {
-        type: LEDGER_ERROR,
-        error: friendlyError,
-      };
-      expect(ledgerError(error)).toEqual(expected);
     });
   });
